frontend: split useAIContext into its own module to satisfy react-refresh
The 'react-refresh/only-export-components' rule rejects any module that
exports a mix of components and non-components (the rule needs the
ability to swap a single component without re-running module
initialization). App.tsx was exporting both <App /> and useAIContext,
so HMR couldn't update App in isolation.

Move the no-op AIContext + provider defaults + useAIContext hook into
src/contexts/AIContext.tsx. App.tsx now only exports the App component.
Update the two consumers (NotebookPage, CellOutputRenderer) to import
from the new path.

Zero behavior change — the context is still a no-op stub kept for
notebook-component compatibility. ESLint should now report 0 warnings.
