+++ frontend/src/components/Notebooks/parts/CellOutputRenderer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DOMPurify from 'dompurify';
 import { Sparkles } from 'lucide-react';
-import { useAIContext } from '@/App';
+import { useAIContext } from '@/contexts/AIContext';
 import { Button } from '@/components/ui/button';
 import { CellOutput } from '@/hooks/useJupyterKernel';
 import { ansiToHtml, stripAnsi } from '../ansiToHtml';
